Reorganization to prepare for deployment
This commit makes a bundle of changes to prepare the server to
be deployed on Cloud Run. This includes:
* A Dockerfile has been added to prepare a Docker image
* Files have been moved into subdirectories for better organization
* A run script has been added to run the server with gunicorn
* Debug run has been improved to automatically create the demo db

Note that FLASK_DEBUG=True should be set to run the software in
debug mode. This should be set in the run configuration to ensure
it is true when the software starts. We cannot set it based on
the flag passed into the app run method, since this creates a
circular dependency if we do.